move directed view test to separate file
